Feature: Enhance Web UI with Navbar, Pages, and Home Redesign
- Added Navbar component with responsive navigation
- Created AboutPage and CommunityPage placeholders
- Redesigned Home page with modern layout and sections
- Updated package.json and pnpm-lock.yaml
- Removed package manager specification in package.json

+++ src/IconGeneratorAI.Persistence/EntityFramework/Configurations/UserBalanceTransactionConfiguration.cs
@@ -29,7 +29,7 @@ public sealed class UserBalanceTransactionConfiguration : IEntityTypeConfigurati
 
         builder.Property(x => x.Type) // 0,1,2 // Add, Remove
         .HasConversion<int>()
-        .HasColumnType("TINYINT") // 0,255
+        .HasColumnType("SMALLINT") // 0,255
         .IsRequired();
 
         // Common Properties
